refactor: Move out test helpers from peg_in.rs for reusability
