+++ src/chainstate/burn/operations/peg_in.rs
@@ -185,9 +185,9 @@ impl From<ClarityRuntimeError> for ParseError {
 
 #[cfg(test)]
 mod tests {
-    use clarity::util::hash::Hash160;
     use rand::rngs::StdRng;
     use rand::SeedableRng;
+    use stacks_common::util::hash::Hash160;
 
     use crate::burnchains::{
         bitcoin::{
@@ -201,24 +201,26 @@ mod tests {
         Txid,
     };
 
+    use crate::chainstate::burn::operations::test_helpers;
+
     use super::*;
 
     #[test]
     fn test_parse_peg_in_should_succeed_given_a_conforming_transaction_without_memo() {
-        let mut rng = seeded_rng();
+        let mut rng = test_helpers::seeded_rng();
         let opcode = Opcodes::PegIn;
 
-        let peg_wallet_address = random_bytes(&mut rng);
+        let peg_wallet_address = test_helpers::random_bytes(&mut rng);
         let amount = 10;
-        let output2 = Output2Data::new_as_option(amount, peg_wallet_address);
+        let output2 = test_helpers::Output2Data::new_as_option(amount, peg_wallet_address);
 
         let mut data = vec![1];
-        let addr_bytes = random_bytes(&mut rng);
+        let addr_bytes = test_helpers::random_bytes(&mut rng);
         let stx_address = StacksAddress::new(1, addr_bytes.into());
         data.extend_from_slice(&addr_bytes);
 
-        let tx = burnchain_transaction(data, output2, opcode);
-        let header = burnchain_block_header();
+        let tx = test_helpers::burnchain_transaction(data, output2, opcode);
+        let header = test_helpers::burnchain_block_header();
 
         let op = PegInOp::from_tx(&header, &tx).expect("Failed to construct peg-in operation");
 
@@ -257,25 +259,25 @@ mod tests {
 
     #[test]
     fn test_parse_peg_in_should_succeed_given_a_contract_recipient() {
-        let mut rng = seeded_rng();
+        let mut rng = test_helpers::seeded_rng();
         let opcode = Opcodes::PegIn;
 
         let contract_name = "This_is_a_valid_contract_name";
-        let peg_wallet_address = random_bytes(&mut rng);
+        let peg_wallet_address = test_helpers::random_bytes(&mut rng);
         let amount = 10;
         let output2 = Output2Data::new_as_option(amount, peg_wallet_address);
-        let memo: [u8; 6] = random_bytes(&mut rng);
+        let memo: [u8; 6] = test_helpers::random_bytes(&mut rng);
 
         let mut data = vec![1];
-        let addr_bytes = random_bytes(&mut rng);
+        let addr_bytes = test_helpers::random_bytes(&mut rng);
         let stx_address = StacksAddress::new(1, addr_bytes.into());
         data.extend_from_slice(&addr_bytes);
         data.extend_from_slice(contract_name.as_bytes());
         data.extend_from_slice(&[0; 11]); // Padding contract name
         data.extend_from_slice(&memo);
 
-        let tx = burnchain_transaction(data, output2, opcode);
-        let header = burnchain_block_header();
+        let tx = test_helpers::burnchain_transaction(data, output2, opcode);
+        let header = test_helpers::burnchain_block_header();
 
         let op = PegInOp::from_tx(&header, &tx).expect("Failed to construct peg-in operation");
 
@@ -320,22 +322,22 @@ mod tests {
 
     #[test]
     fn test_parse_peg_in_should_return_error_given_wrong_opcode() {
-        let mut rng = seeded_rng();
+        let mut rng = test_helpers::seeded_rng();
         let opcode = Opcodes::StackStx;
 
-        let peg_wallet_address = random_bytes(&mut rng);
+        let peg_wallet_address = test_helpers::random_bytes(&mut rng);
         let amount = 10;
         let output2 = Output2Data::new_as_option(amount, peg_wallet_address);
-        let memo: [u8; 6] = random_bytes(&mut rng);
+        let memo: [u8; 6] = test_helpers::random_bytes(&mut rng);
 
         let mut data = vec![1];
-        let addr_bytes: [u8; 20] = random_bytes(&mut rng);
+        let addr_bytes: [u8; 20] = test_helpers::random_bytes(&mut rng);
         data.extend_from_slice(&addr_bytes);
         data.extend_from_slice(&[0; 40]); // Padding contract name
         data.extend_from_slice(&memo);
 
-        let tx = burnchain_transaction(data, output2, opcode);
-        let header = burnchain_block_header();
+        let tx = test_helpers::burnchain_transaction(data, output2, opcode);
+        let header = test_helpers::burnchain_block_header();
 
         let op = PegInOp::from_tx(&header, &tx);
 
@@ -349,23 +351,23 @@ mod tests {
     fn test_parse_peg_in_should_return_error_given_invalid_utf8_contract_name() {
         let invalid_utf8_byte_sequence = [255, 255];
 
-        let mut rng = seeded_rng();
+        let mut rng = test_helpers::seeded_rng();
         let opcode = Opcodes::PegIn;
 
-        let peg_wallet_address = random_bytes(&mut rng);
+        let peg_wallet_address = test_helpers::random_bytes(&mut rng);
         let amount = 10;
         let output2 = Output2Data::new_as_option(amount, peg_wallet_address);
-        let memo: [u8; 6] = random_bytes(&mut rng);
+        let memo: [u8; 6] = test_helpers::random_bytes(&mut rng);
 
         let mut data = vec![1];
-        let addr_bytes: [u8; 20] = random_bytes(&mut rng);
+        let addr_bytes: [u8; 20] = test_helpers::random_bytes(&mut rng);
         data.extend_from_slice(&addr_bytes);
         data.extend_from_slice(&invalid_utf8_byte_sequence);
         data.extend_from_slice(&[0; 40]); // Padding contract name
         data.extend_from_slice(&memo);
 
-        let tx = burnchain_transaction(data, output2, opcode);
-        let header = burnchain_block_header();
+        let tx = test_helpers::burnchain_transaction(data, output2, opcode);
+        let header = test_helpers::burnchain_block_header();
 
         let op = PegInOp::from_tx(&header, &tx);
 
@@ -377,19 +379,19 @@ mod tests {
 
     #[test]
     fn test_parse_peg_in_should_return_error_given_no_second_output() {
-        let mut rng = seeded_rng();
+        let mut rng = test_helpers::seeded_rng();
         let opcode = Opcodes::PegIn;
 
         let memo: [u8; 6] = random_bytes(&mut rng);
 
         let mut data = vec![1];
-        let addr_bytes: [u8; 20] = random_bytes(&mut rng);
+        let addr_bytes: [u8; 20] = test_helpers::random_bytes(&mut rng);
         data.extend_from_slice(&addr_bytes);
         data.extend_from_slice(&[0; 40]); // Padding contract name
         data.extend_from_slice(&memo);
 
-        let tx = burnchain_transaction(data, None, opcode);
-        let header = burnchain_block_header();
+        let tx = test_helpers::burnchain_transaction(data, None, opcode);
+        let header = test_helpers::burnchain_block_header();
 
         let op = PegInOp::from_tx(&header, &tx);
 
@@ -401,19 +403,19 @@ mod tests {
 
     #[test]
     fn test_parse_peg_in_should_return_error_given_too_short_data_array() {
-        let mut rng = seeded_rng();
+        let mut rng = test_helpers::seeded_rng();
         let opcode = Opcodes::PegIn;
 
-        let peg_wallet_address = random_bytes(&mut rng);
+        let peg_wallet_address = test_helpers::random_bytes(&mut rng);
         let amount = 10;
-        let output2 = Output2Data::new_as_option(amount, peg_wallet_address);
+        let output2 = test_helpers::Output2Data::new_as_option(amount, peg_wallet_address);
 
         let mut data = vec![1];
-        let addr_bytes: [u8; 19] = random_bytes(&mut rng);
+        let addr_bytes: [u8; 19] = test_helpers::random_bytes(&mut rng);
         data.extend_from_slice(&addr_bytes);
 
-        let tx = burnchain_transaction(data, output2, opcode);
-        let header = burnchain_block_header();
+        let tx = test_helpers::burnchain_transaction(data, output2, opcode);
+        let header = test_helpers::burnchain_block_header();
 
         let op = PegInOp::from_tx(&header, &tx);
 
@@ -425,13 +427,13 @@ mod tests {
 
     #[test]
     fn test_check_should_return_error_on_zero_amount_and_ok_on_any_other_values() {
-        let mut rng = seeded_rng();
+        let mut rng = test_helpers::seeded_rng();
 
         let peg_wallet_address = random_bytes(&mut rng);
-        let memo: [u8; 6] = random_bytes(&mut rng);
+        let memo: [u8; 6] = test_helpers::random_bytes(&mut rng);
 
         let mut data = vec![1];
-        let addr_bytes = random_bytes(&mut rng);
+        let addr_bytes = test_helpers::random_bytes(&mut rng);
         let stx_address = StacksAddress::new(1, addr_bytes.into());
         data.extend_from_slice(&addr_bytes);
         data.extend_from_slice(&[0; 40]); // Padding contract name
@@ -439,10 +441,11 @@ mod tests {
 
         let create_op = move |amount| {
             let opcode = Opcodes::PegIn;
-            let output2 = Output2Data::new_as_option(amount, peg_wallet_address.clone());
+            let output2 =
+                test_helpers::Output2Data::new_as_option(amount, peg_wallet_address.clone());
 
-            let tx = burnchain_transaction(data.clone(), output2, opcode);
-            let header = burnchain_block_header();
+            let tx = test_helpers::burnchain_transaction(data.clone(), output2, opcode);
+            let header = test_helpers::burnchain_block_header();
 
             PegInOp::from_tx(&header, &tx).expect("Failed to construct peg-in operation")
         };
@@ -462,79 +465,5 @@ mod tests {
         create_op(u64::MAX)
             .check()
             .expect("Any strictly positive amounts should be ok");
-    }
-
-    fn seeded_rng() -> StdRng {
-        SeedableRng::from_seed([0; 32])
-    }
-
-    fn random_bytes<Rng: rand::Rng, const N: usize>(rng: &mut Rng) -> [u8; N] {
-        [rng.gen(); N]
-    }
-
-    fn burnchain_block_header() -> BurnchainBlockHeader {
-        BurnchainBlockHeader {
-            block_height: 0,
-            block_hash: [0; 32].into(),
-            parent_block_hash: [0; 32].into(),
-            num_txs: 0,
-            timestamp: 0,
-        }
-    }
-
-    fn burnchain_transaction(
-        data: Vec<u8>,
-        output2: Option<Output2Data>,
-        opcode: Opcodes,
-    ) -> BurnchainTransaction {
-        BurnchainTransaction::Bitcoin(bitcoin_transaction(data, output2, opcode))
-    }
-
-    fn bitcoin_transaction(
-        data: Vec<u8>,
-        output2: Option<Output2Data>,
-        opcode: Opcodes,
-    ) -> BitcoinTransaction {
-        BitcoinTransaction {
-            txid: Txid([0; 32]),
-            vtxindex: 0,
-            opcode: opcode as u8,
-            data,
-            data_amt: 0,
-            inputs: vec![BitcoinTxInputStructured {
-                keys: vec![],
-                num_required: 0,
-                in_type: BitcoinInputType::Standard,
-                tx_ref: (Txid([0; 32]), 0),
-            }
-            .into()],
-            outputs: output2
-                .into_iter()
-                .map(|output2data| output2data.as_bitcoin_tx_output())
-                .collect(),
-        }
-    }
-
-    struct Output2Data {
-        amount: u64,
-        peg_wallet_address: [u8; 32],
-    }
-
-    impl Output2Data {
-        fn new_as_option(amount: u64, peg_wallet_address: [u8; 32]) -> Option<Self> {
-            Some(Self {
-                amount,
-                peg_wallet_address,
-            })
-        }
-        fn as_bitcoin_tx_output(&self) -> BitcoinTxOutput {
-            BitcoinTxOutput {
-                units: self.amount,
-                address: BitcoinAddress::Segwit(SegwitBitcoinAddress::P2TR(
-                    true,
-                    self.peg_wallet_address,
-                )),
-            }
-        }
     }
 }
