+++ src/chainstate/burn/operations/test_helpers.rs
@@ -1,0 +1,88 @@
+use rand::rngs::StdRng;
+use rand::SeedableRng;
+use stacks_common::util::hash::Hash160;
+
+use crate::burnchains::BurnchainBlockHeader;
+use crate::burnchains::BurnchainTransaction;
+use crate::burnchains::{
+    bitcoin::{
+        address::{
+            BitcoinAddress, LegacyBitcoinAddress, LegacyBitcoinAddressType, SegwitBitcoinAddress,
+        },
+        BitcoinInputType, BitcoinNetworkType, BitcoinTransaction, BitcoinTxInputStructured,
+        BitcoinTxOutput,
+    },
+    Txid,
+};
+use crate::chainstate::burn::Opcodes;
+
+pub(crate) fn seeded_rng() -> StdRng {
+    SeedableRng::from_seed([0; 32])
+}
+
+pub(crate) fn random_bytes<Rng: rand::Rng, const N: usize>(rng: &mut Rng) -> [u8; N] {
+    [rng.gen(); N]
+}
+
+pub(crate) fn burnchain_block_header() -> BurnchainBlockHeader {
+    BurnchainBlockHeader {
+        block_height: 0,
+        block_hash: [0; 32].into(),
+        parent_block_hash: [0; 32].into(),
+        num_txs: 0,
+        timestamp: 0,
+    }
+}
+
+pub(crate) fn burnchain_transaction(
+    data: Vec<u8>,
+    output2: Option<Output2Data>,
+    opcode: Opcodes,
+) -> BurnchainTransaction {
+    BurnchainTransaction::Bitcoin(bitcoin_transaction(data, output2, opcode))
+}
+
+fn bitcoin_transaction(
+    data: Vec<u8>,
+    output2: Option<Output2Data>,
+    opcode: Opcodes,
+) -> BitcoinTransaction {
+    BitcoinTransaction {
+        txid: Txid([0; 32]),
+        vtxindex: 0,
+        opcode: opcode as u8,
+        data,
+        data_amt: 0,
+        inputs: vec![BitcoinTxInputStructured {
+            keys: vec![],
+            num_required: 0,
+            in_type: BitcoinInputType::Standard,
+            tx_ref: (Txid([0; 32]), 0),
+        }
+        .into()],
+        outputs: output2
+            .into_iter()
+            .map(|output2data| output2data.as_bitcoin_tx_output())
+            .collect(),
+    }
+}
+
+pub(crate) struct Output2Data {
+    amount: u64,
+    address: [u8; 32],
+}
+
+impl Output2Data {
+    pub(crate) fn new_as_option(amount: u64, peg_wallet_address: [u8; 32]) -> Option<Self> {
+        Some(Self {
+            amount,
+            address: peg_wallet_address,
+        })
+    }
+    pub(crate) fn as_bitcoin_tx_output(&self) -> BitcoinTxOutput {
+        BitcoinTxOutput {
+            units: self.amount,
+            address: BitcoinAddress::Segwit(SegwitBitcoinAddress::P2TR(true, self.address)),
+        }
+    }
+}
